+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added the default binding `_choice`, which returns the contents of the
   last chosen choice
 - Added the `-=`, `+=`, `/=`, and `*=` assignment operators
+- Filter operator `|` support
 
 ### Changed
 - The `OnClear` event in RumorState has been changed to use a `ClearType` enum,